// src/Modelos/Usuario.d.ts
export type Usuario = {
    Id: string;
    Admin_Id?: string;
    Name: string;
    LastName: string;
    Email: string;
    State: string;
    City: string;
    Role_Id: string;
    Password: string;
    Is_Verified: boolean;
    Insignia: string;
    Token: string;
    User: string;
    Service?: string;
    Occupation?: string;
    Account?: string;
    AccNumber?: string;
    Bank?: string;
    English?: boolean;
    Notification?: boolean;
}